refactor(language-core): extract binding-reference analysis into bindingReferences.ts
